pyproject.toml
[project]
name = "chat-lm"
version = "0.1.0"
description = "Train my own llm to response to chat messages"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.6.1",
    "markupsafe==2.1.3",
    "matplotlib>=3.10.8",
    "pyyaml>=6.0.3",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "wandb>=0.25.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"
